fdlog and xastir
fdlog apparently needs to be exicuted with python instead of
./fdlog_scicsg.py

xastir update: apparently if the xastir binary is run first before the
uninstall and compile/install from source the configuration in the user
directory will point to the binary's location of the files. added code
to check for the file and move it into a backup. on run xastir should
re-make the config file.